requirements.txt
Django==5.0.2
djangorestframework==3.15.2
python-dotenv==1.1.1
Pillow==10.2.0
gunicorn==21.2.0
django-cleanup==8.1.0
django-tailwind-cli==2.1.0
django-csp==3.8
whitenoise==6.7.0
django-storages==1.14.4
dj-database-url==2.3.0
psycopg2-binary==2.9.9
django-debug-toolbar==4.4.6

# Code Quality & Testing
flake8==7.1.1
black==24.8.0
isort==5.13.2
coverage==7.6.1
factory-boy==3.3.3

# API Enhancements
drf-spectacular==0.27.2
django-anymail[sendgrid]==14.0